Add allocator to Array and List

+++ include/riku/func.h
@@ -35,8 +35,8 @@ public:
 
             struct
             {
-                usize size;
-                byte  functor[sizeof(FuncType)];
+                int  size;
+                byte functor[sizeof(FuncType)];
             };
 
             struct
